refactor(plan): Pull out basics
